@@ -1,17 +1,24 @@
-cmake_minimum_required(VERSION 3.17)
-project(Indexer)
+CMAKE_MINIMUM_REQUIRED(VERSION 3.13.4)
+PROJECT(Indexer)
 
-find_package(Boost 1.54.0 COMPONENTS system filesystem REQUIRED)
+SET(CMAKE_CXX_STANDARD 17)
+#SET(CMAKE_CXX_FLAGS "${CMAKE_C_FLAGS} -Wall -Wextra -Werror -Wpedantic")
 
-set(CMAKE_CXX_STANDARD 14)
+FIND_PACKAGE(Boost COMPONENTS system filesystem REQUIRED)
+INCLUDE_DIRECTORIES(${Boost_INCLUDE_DIR})
 
-INCLUDE_DIRECTORIES(${PROJECT_SOURCE_DIR}/../utils)
+INCLUDE_DIRECTORIES(../../utils/include)
+INCLUDE_DIRECTORIES(../utils)
+
+FIND_LIBRARY(InternalDB-dev REQUIRED)
+INCLUDE_DIRECTORIES(../../DataBases/InternalDB/include)
+
 INCLUDE_DIRECTORIES(include)
-
-add_executable(Indexer Indexer.cpp)
-
-ADD_LIBRARY(libIndexer STATIC
+ADD_LIBRARY(${PROJECT_NAME}-dev SHARED
         include/Indexer.h
-        Indexer.cpp
-        ../utils/File.h
-        ../utils/File.cpp)
+        src/Indexer.cpp
+        )
+TARGET_LINK_LIBRARIES(${PROJECT_NAME}-dev InternalDB-dev ${Boost_LIBRARIES})
+#add_executable(Indexer Indexer.cpp)
+
+
