@@ -1,24 +1,33 @@
 #pragma once
 
+#define CHUNK_SIZE 4096
+
+#include "structs/Chunk.h"
 #include "File.h"
 #include <openssl/md5.h>
 #include <iomanip>
 #include "zlib.h"
 #include <memory>
+#include <algorithm>
+#include <iterator>
+#include <sstream>
+#include <array>
+
 
 class Chunker {
  public:
-  explicit Chunker(File);
+  explicit Chunker(const File&);
   int SentNewChunk();
   int SentNewPosition();
   File *GetFile();
-  std::vector<FileChunk> ChunkFile();
-  void MergeFile(std::vector<FileChunk>);
-  void ChunkCompare(FileChunk data);
+  std::vector<Chunk> ChunkFile();
+  std::vector<Chunk> UpdateChunkFile(const std::vector<Chunk>&);
+  void MergeFile(std::vector<Chunk>);
+  void ChunkCompare(Chunk data);
  private:
   File _file;
-  void getRHash(FileChunk &);
-  void getSHash(FileChunk &);
+  std::string getRHash(char*, int);
+  std::string getSHash(char*, int);
   std::string getOldCheckSums();
 
 };
