@@ -1,13 +1,11 @@
 #pragma once
-#include "stdlib.h"
 #include <vector>
 #include <string>
-#include "FileChunk.h"
 #include <fstream>
 
 class File {
  public:
-  File(std::string);
+  explicit File(std::string);
   ~File();
   std::ifstream Read();
   std::ofstream Write();
